+++ frontend/src/App.tsx
@@ -3,13 +3,19 @@ import { QueryClientProvider } from "./query";
 import { RouterProvider } from "./router";
 import { ThemeProvider } from "./context/ThemeContext";
 
+import { ErrorBoundary } from "./components/common/ErrorBoundary";
+import { SkipLink } from "./components/common/SkipLink";
+
 function App() {
   return (
-    <ThemeProvider>
-      <QueryClientProvider>
-        <RouterProvider />
-      </QueryClientProvider>
-    </ThemeProvider>
+    <ErrorBoundary>
+      <SkipLink />
+      <ThemeProvider>
+        <QueryClientProvider>
+          <RouterProvider />
+        </QueryClientProvider>
+      </ThemeProvider>
+    </ErrorBoundary>
   );
 }